add feign ratingClient in driver service

+++ rating-service/src/main/java/com/example/rating_service/model/entity/UserRatingEntity.java
@@ -29,5 +29,5 @@ public class UserRatingEntity {
 
     private Long ratingCount;
 
-    private BigDecimal ratingSum;
+    private Long ratingSum;
 }
